Incorporate CpuPool into the executor crate as ThreadPool, revamp it
