--- futures-cpupool/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "futures-cpupool"
-version = "0.2.0"
-authors = ["Alex Crichton <alex@alexcrichton.com>"]
-license = "MIT/Apache-2.0"
-repository = "https://github.com/alexcrichton/futures-rs"
-homepage = "https://github.com/alexcrichton/futures-rs"
-documentation = "https://docs.rs/futures-cpupool"
-description = """
-An implementation of thread pools which hand out futures to the results of the
-computation on the threads themselves.
-"""
-
-[dependencies]
-num_cpus = "1.0"
-futures = { path = "../futures", version = "0.2", default-features = false }
-futures-executor = { path = "../futures-executor", version = "0.2", default-features = false }
-
-[features]
-std = ["futures/std"]
-default = ["std"]
